🚚 [projects] Move function `updateproject` from `services` to `update` module
